@@ -6,7 +6,7 @@
   "scripts": {
     "build": "babel src --out-dir lib --copy-files",
     "build:watch": "npm run build -- --watch",
-    "test": "NODE_ENV=test mocha --compilers js:babel-core/register --reporter spec --check-leaks",
+    "test": "NODE_ENV=test mocha --require babel-core/register --reporter spec --check-leaks",
     "prepare": "npm run build"
   },
   "keywords": [
@@ -23,28 +23,29 @@
     "url": "https://github.com/Atyantik/pwa-srcset-loader"
   },
   "dependencies": {
-    "datauri": "1.0.5",
-    "get-pixels": "3.3.0",
-    "image-size": "0.5.1",
+    "datauri": "^1.1.0",
+    "get-pixels": "^3.3.2",
+    "image-size": "^0.6.3",
     "loader-utils": "1.1.0",
     "mime": "^2.0.3",
-    "sharp": "^0.18.2"
+    "sharp": "^0.20.8"
   },
   "devDependencies": {
     "babel-cli": "^6.24.0",
     "babel-preset-es2015": "^6.24.0",
-    "chai": "^3.5.0",
-    "chai-spies": "^0.7.1",
-    "dirty-chai": "^1.2.2",
-    "eslint": "^3.16.1",
-    "eslint-config-airbnb-base": "11.1.2",
-    "eslint-plugin-import": "2.2.0",
-    "file-loader": "^1.1.4",
-    "image-webpack-loader": "3.3.0",
-    "jsdom": "^9.12.0",
+    "canvas": "^1.6.12",
+    "chai": "^4.2.0",
+    "chai-spies": "^1.0.0",
+    "dirty-chai": "^2.0.1",
+    "eslint": "^5.6.0",
+    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-plugin-import": "^2.14.0",
+    "file-loader": "^2.0.0",
+    "image-webpack-loader": "^4.3.1",
+    "jsdom": "^12.0.0",
     "lodash": "^4.17.4",
     "memory-fs": "^0.4.1",
-    "mocha": "^3.2.0",
-    "webpack": "2.3.2"
+    "mocha": "^5.2.0",
+    "webpack": "^4.20.2"
   }
 }
