@@ -13,21 +13,21 @@ export function buildSrcSet(img) {
 
 export function blurPlaceholder(img, toBase64) {
   const placeholder = img.placeholder || img;
-  const ratio = placeholder.ratio;
+  const { ratio } = placeholder;
 
   const svg = `
-<svg 
+<svg
   xmlns:xlink="http://www.w3.org/1999/xlink"
-  xmlns="http://www.w3.org/2000/svg" 
+  xmlns="http://www.w3.org/2000/svg"
   viewBox="0 0 ${10 * ratio} 10"
 >
   <filter id="x">
     <feGaussianBlur stdDeviation="1" />
   </filter>
   <image
-    width="100%" 
-    height="100%"  
-    xlink:href="${placeholder.url}" 
+    width="100%"
+    height="100%"
+    xlink:href="${placeholder.url}"
     filter="url(#x)"
   />
 </svg>
